src/requirements.txt
customtkinter==5.0.2
darkdetect==0.7.1
greenlet==2.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.1
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
python-dateutil==2.8.2
pytz==2022.6
shiboken6==6.4.1
six==1.16.0
SQLAlchemy==1.4.44
